refactor(librarytest): 필요없는 패키징 제거
