+++ src/test/kotlin/library/subdividerule/ForbiddenFourTest.kt
@@ -1,4 +1,4 @@
-package library.subdividerule.forbiddenfour
+package library.subdividerule
 
 import domain.CoordinateState
 import domain.Position
